Update peer review with reviewing team and document in the title
It would be nice for the marking TA to be able to easily count the number of reviews by a team and know which document it is for. This is particularly needed if multiple teams review a document, and now for deliverables with multiple documents due at once (e.g. SRS/HA).